move demo filter
