+++ examples/swift/hello_world/ViewController.swift
@@ -6,29 +6,6 @@ private let kRequestAuthority = "api.lyft.com"
 private let kRequestPath = "/ping"
 private let kRequestScheme = "https"
 
-struct PlatformDemoFilter: ResponseFilter {
-  let name = "PlatformStub"
-
-  func onResponseHeaders(_ headers: ResponseHeaders, endStream: Bool)
-    -> FilterHeaderStatus<ResponseHeaders>
-  {
-    NSLog("Adding new header!")
-    let builder = headers.toResponseHeadersBuilder()
-    builder.add(name: "filter-demo", value: "1")
-    return .continue(builder.build())
-  }
-
-  func setResponseFilterCallbacks(_ callbacks: ResponseFilterCallbacks) {}
-
-  func onResponseData(_ body: Data, endStream: Bool) -> FilterDataStatus {
-    return .continue(body)
-  }
-
-  func onResponseTrailers(_ trailers: ResponseTrailers) -> FilterTrailerStatus<ResponseTrailers> {
-    return .continue(trailers)
-  }
-}
-
 final class ViewController: UITableViewController {
   private var results = [Result<Response, RequestError>]()
   private var timer: Timer?
@@ -38,7 +15,7 @@ final class ViewController: UITableViewController {
     super.viewDidLoad()
     do {
       NSLog("starting Envoy...")
-      self.client = try StreamClientBuilder().addFilter(PlatformDemoFilter()).build()
+      self.client = try StreamClientBuilder().addFilter(DemoFilter()).build()
     } catch let error {
       NSLog("starting Envoy failed: \(error)")
     }
